feat(tokenManager.js): moved tokenManager into it's own file and changed form to ES
