+++ manifest.json
@@ -21,7 +21,8 @@
     "https://calendar.google.com/"
   ],
   "background": {
-    "service_worker": "background.js"
+    "service_worker": "background.js",
+    "type": "module"
   },
   "action": {
     "default_popup": "popup.html"
